Changing h and cpp for beter output

+++ import/Manifest.h
@@ -9,7 +9,7 @@ class Manifest : public Inspection
 private:
 	ShippingUnit *root;
 	vector<ShippingUnit *> list;
-	int currentIndex;
+	size_t currentIndex;
 	void populateList(ShippingUnit *unit);
 
 public:
